[refact] Change the source file to header for refactoring
