+++ Shell/Shell.vcxproj
@@ -132,30 +132,29 @@
     <ClCompile Include="config.cpp" />
     <ClCompile Include="logger.cpp" />
     <ClCompile Include="logger_exception.cpp" />
-    <ClCompile Include="script_loader.cpp" />
-    <ClCompile Include="script_parser.cpp" />
+    <ClCompile Include="script_loader.h" />
+    <ClCompile Include="script_parser.h" />
     <ClCompile Include="mock_ssddriver.h" />
-    <ClCompile Include="script_store.cpp" />
-    <ClCompile Include="shell.cpp" />
-    <ClCompile Include="shell_command_factory.cpp" />
-    <ClCompile Include="shell_erase_command.cpp" />
-    <ClCompile Include="shell_erase_range_command.cpp" />
-    <ClCompile Include="shell_exception.cpp" />
-    <ClCompile Include="shell_exit_command.cpp" />
-    <ClCompile Include="shell_flush_command.cpp" />
-    <ClCompile Include="shell_full_read_command.cpp" />
-    <ClCompile Include="shell_full_write_command.cpp" />
-    <ClCompile Include="shell_help_command.cpp" />
-    <ClCompile Include="shell_script_compare_command.cpp" />
-    <ClCompile Include="shell_script_loop_command.cpp" />
-    <ClCompile Include="shell_script_run_command.cpp" />
-    <ClCompile Include="shell_script_write_command.cpp" />
-    <ClCompile Include="shell_util.cpp" />
-    <ClCompile Include="shell_read_command.cpp" />
-    <ClCompile Include="shell_write_command.cpp" />
+    <ClCompile Include="script_store.h" />
+    <ClCompile Include="shell.h" />
+    <ClCompile Include="shell_command_factory.h" />
+    <ClCompile Include="shell_erase_command.h" />
+    <ClCompile Include="shell_erase_range_command.h" />
+    <ClCompile Include="shell_exception.h" />
+    <ClCompile Include="shell_executor.h" />
+    <ClCompile Include="shell_exit_command.h" />
+    <ClCompile Include="shell_flush_command.h" />
+    <ClCompile Include="shell_full_read_command.h" />
+    <ClCompile Include="shell_full_write_command.h" />
+    <ClCompile Include="shell_help_command.h" />
+    <ClCompile Include="shell_script_compare_command.h" />
+    <ClCompile Include="shell_script_loop_command.h" />
+    <ClCompile Include="shell_script_run_command.h" />
+    <ClCompile Include="shell_script_write_command.h" />
+    <ClCompile Include="shell_read_command.h" />
+    <ClCompile Include="shell_write_command.h" />
     <ClCompile Include="ssddriver_interface.h" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="shell_executor.cpp" />
     <ClCompile Include="shell_interface_ut.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
@@ -179,6 +178,7 @@
     <ClInclude Include="script_parameter_generator.h" />
     <ClInclude Include="shell_command_interface.h" />
     <ClInclude Include="shell_script_command_interface.h" />
+    <ClInclude Include="shell_util.h" />
     <ClInclude Include="ssddriver.h" />
   </ItemGroup>
   <ItemGroup>
