+++ Shell/main.cpp
@@ -9,7 +9,7 @@ int main(int argc, char** argv) {
     return RUN_ALL_TESTS();
 }
 #else
-#include "shell.cpp"
+#include "shell.h"
 #include "logger.h"
 
 int main(int argc, char** argv) {
