+++ Shell/shell_erase_range_command.h
@@ -1,21 +1,21 @@
 #pragma once
 #include "shell_command_interface.h"
-#include "shell_exception.cpp"
-#include "shell_util.cpp"
+#include "shell_exception.h"
+#include "shell_util.h"
 #include "ssddriver_interface.h"
 
-class ShellWriteCommand : public ShellCommandInterface {
+class ShellEraseRangeCommand : public ShellCommandInterface {
 public:
-	ShellWriteCommand(SsdDriverInterface* pDriverInterface) {
+	ShellEraseRangeCommand(SsdDriverInterface* pDriverInterface) {
 		mpDriverInterface = pDriverInterface;
 	}
 
 	string execute(vector<string> args) {
 		try {
 			vector<unsigned int> convertedArgs = convertCmdArgs(args);
-			string output = "[Write] Done";
+			string output = "[Full Erase] Done";
 
-			mpDriverInterface->writeSSD((int)convertedArgs[0], convertedArgs[1]);
+			mpDriverInterface->eraseSSD((int)convertedArgs[0], (int)convertedArgs[1]);
 
 			return output;
 		}
@@ -30,12 +30,14 @@ private:
 	vector<unsigned int>  convertCmdArgs(vector<string> args) {
 		try {
 			vector<unsigned int> output;
+
 			if (args.size() != 3) {
 				throw ShellArgConvertException("args parameter size invalid");
 			}
 
+			// LBA String으로 변환.
 			output.push_back(ShellUtil::getUtilObj().convertDecimalStringForLba(args[1]));
-			output.push_back(ShellUtil::getUtilObj().convertHexStringForData(args[2]));
+			output.push_back(ShellUtil::getUtilObj().convertLBAtoSize(args[1], args[2]));
 
 			return output;
 		}
