+++ Shell/shell_exception.h
