+++ Shell/shell_exit_command.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "shell_command_interface.h"
-#include "shell_exception.cpp"
+#include "shell_exception.h"
 
 class ShellExitCommand : public ShellCommandInterface {
 public:
