+++ Shell/shell_full_write_command.h
@@ -1,23 +1,26 @@
 #pragma once
 #include "shell_command_interface.h"
-#include "shell_exception.cpp"
-#include "shell_util.cpp"
+#include "shell_exception.h"
+#include "shell_util.h"
 #include "ssddriver_interface.h"
 
-class ShellEraseRangeCommand : public ShellCommandInterface {
+class ShellFullWriteCommand : public ShellCommandInterface {
 public:
-	ShellEraseRangeCommand(SsdDriverInterface* pDriverInterface) {
+	ShellFullWriteCommand(SsdDriverInterface* pDriverInterface) {
 		mpDriverInterface = pDriverInterface;
 	}
 
 	string execute(vector<string> args) {
 		try {
 			vector<unsigned int> convertedArgs = convertCmdArgs(args);
-			string output = "[Full Erase] Done";
+			string fullWriteResult = "[Full Write] Done";
 
-			mpDriverInterface->eraseSSD((int)convertedArgs[0], (int)convertedArgs[1]);
+			for (int i = 0; i < 100; i++) {
 
-			return output;
+				mpDriverInterface->writeSSD(i, convertedArgs[0]);
+			}
+
+			return fullWriteResult;
 		}
 		catch (ShellArgConvertException e) {
 			throw e;
@@ -30,14 +33,11 @@ private:
 	vector<unsigned int>  convertCmdArgs(vector<string> args) {
 		try {
 			vector<unsigned int> output;
-
-			if (args.size() != 3) {
+			if (args.size() != 2) {
 				throw ShellArgConvertException("args parameter size invalid");
 			}
 
-			// LBA String으로 변환.
-			output.push_back(ShellUtil::getUtilObj().convertDecimalStringForLba(args[1]));
-			output.push_back(ShellUtil::getUtilObj().convertLBAtoSize(args[1], args[2]));
+			output.push_back(ShellUtil::getUtilObj().convertHexStringForData(args[1]));
 
 			return output;
 		}
