+++ Shell/shell_help_command.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "shell_command_interface.h"
-#include "shell_exception.cpp"
-#include "script_store.cpp"
+#include "shell_exception.h"
+#include "script_store.h"
 
 class ShellHelpCommand : public ShellCommandInterface {
 public:
