+++ Shell/shell_interface_ut.cpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "shell_executor.cpp"
+#include "shell_executor.h"
 #include "mock_ssddriver.h"
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
