+++ Shell/shell_script_run_command.h
@@ -1,9 +1,9 @@
 #pragma once
-#include "shell_exception.cpp"
-#include "shell_util.cpp"
+#include "shell_exception.h"
+#include "shell_util.h"
 #include "shell_command_interface.h"
 #include "shell_script_command_interface.h"
-#include "script_store.cpp"
+#include "script_store.h"
 #include "script_parameter_generator.h"
 #include <memory>
